Added: Workshop-1

+++ src/com/bridgelabz/workshop/QuotientAndRemainder.java
@@ -18,8 +18,8 @@ public class QuotientAndRemainder {
         System.out.println("Enter B:");
         Double b = sc.nextDouble();
 
-        double quotient = (a/b);
-        double remainder = (a%b);
+        double quotient = (int)(a/b);
+        double remainder = (int)(a%b);
 
         System.out.println(a+" (A) when divided by "+b+" (B) gives the quotient "+quotient+" and the remainder "+remainder);
